+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   will now be suggested
   (for [QOwnNotesAPI Issue #11](https://github.com/pbek/qownnotesapi/issues/11))
     - the placeholder text was updated to be more informative
+- broken links between notes and tags are now automatically removed when the 
+  tag tree is reloaded (for [Issue #548](https://github.com/pbek/QOwnNotes/issues/548))
 
 ## 17.03.5
 - fixed more url highlighting and clicking issues